+++ src/arch/x86_32/fb.h
@@ -1,15 +1,14 @@
 #ifndef X86_FB_HEADER
 #define X86_FB_HEADER
 
-#include <stdint.h>
-#include <stddef.h>
+#include <library.h>
 
-static const size_t FB_WIDTH = 80;
-static const size_t FB_HEIGHT = 25;
-static const uint16_t FB_COMMAND_PORT = 0x3D4;
-static const uint16_t FB_DATA_PORT = 0x3D5;
-static const uint8_t FB_HIGH_BYTE_COMMAND = 14;
-static const uint8_t FB_LOW_BYTE_COMMAND = 15;
+static const u32 FB_WIDTH = 80;
+static const u32 FB_HEIGHT = 25;
+static const u16 FB_COMMAND_PORT = 0x3D4;
+static const u16 FB_DATA_PORT = 0x3D5;
+static const u8 FB_HIGH_BYTE_COMMAND = 14;
+static const u8 FB_LOW_BYTE_COMMAND = 15;
 
 typedef enum fb_color_t_enum {
 	FB_COLOR_BLACK = 0,
@@ -32,14 +31,14 @@ typedef enum fb_color_t_enum {
  
 void fb_initialize();
 void fb_set_color(fb_color_t fg, fb_color_t bg);
-void fb_cursor(size_t x, size_t y);
-void fb_place_char(char c, fb_color_t color, size_t x, size_t y);
+void fb_cursor(u32 x, u32 y);
+void fb_place_char(char c, fb_color_t color, u32 x, u32 y);
 void fb_print_char(char c);
-void fb_print_nstring(const char * string, size_t size);
+void fb_print_nstring(const char * string, u32 size);
 void fb_print_string(const char * string);
-void fb_print_uint(uint32_t value);
-void fb_print_int(int32_t value);
-void fb_print_hex(uint32_t value);
-void fb_print_byte(uint8_t value);
+void fb_print_uint(u32 value);
+void fb_print_int(i32 value);
+void fb_print_hex(u32 value);
+void fb_print_byte(u8 value);
 
 #endif
